Added redirect to paypal on event registration
Also fixed history, eg browsing to a page that requires login, after login you will be brought to the requested page instead of home.

Wrapped getAccessTokenSilently() in try catch block, login with redirect if this raises an exception. This should fix an issue whereby the app would crash if your token expired and you havent logged back in yet.